# back-ogs/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OGS",
        "contact": {
            "email": "mickpicart@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "OGS API"
        }
    ],
    "paths": {
...
        "/loadextdataserrors/{id}": {
            "put": {
                "tags": [
                    "External Datas and Errors by Id"
                ],
                "summary": "Load external datas and errors to DataBase for one website",
                "description": "Load external datas and errors to DataBase for one website",
                "operationId": "loadOneWebsiteExtDatasAndErrorsToDb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The website id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
        "/website": {
            "get": {
                "tags": [
                    "Websites"
                ],
                "summary": "Get websites list",
                "description": "Returns websites list from DataBase",
                "operationId": "websitesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Websites"
                ],
                "summary": "Register a new website",
                "operationId": "websiteRegister",
                "parameters": [
                    {
                        "name": "url",
                        "in": "query",
                        "description": "The website url",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Website successfully registered",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "409": {
                        "description": "Resource already exists"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/website/{id}": {
            "delete": {
                "tags": [
                    "Websites"
                ],
                "summary": "Delete a website and all its datas from Database using its id",
                "operationId": "websiteSupervised",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The website id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Website and related datas successfully deleted from Database",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Websites"
                ],
                "summary": "Modify a website supervision status",
                "operationId": "websiteSupervised",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The website id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Website supervision parameter successfully modified",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "access_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}